everything but specials work

+++ main.py
@@ -62,32 +62,32 @@ def menu():
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
     cursor.execute("SELECT DISTINCT itemName, itemPrice FROM Menu WHERE itemCategory = 'Pizza' ORDER BY itemName")
     menu_items = cursor.fetchall()
+    if is_admin():
     # this is for deleting the menu items
-    if request.method == 'POST' and 'delete_menu' in request.form:
+        if request.method == 'POST' and 'delete_menu' in request.form:
             menu_id = request.form['delete_menu']
             cursor.execute('DELETE FROM menu WHERE itemName = %s', (menu_id,))
             mysql.connection.commit()
-    if request.method == 'POST' and 'add-item' in request.form and 'add-price' in request.form and 'add-category' in request.form:
-            item_name = request.form['add-item']
-            item_price = request.form['add-price']
-            item_category = request.form['add-category']
+        if request.method == 'POST' and 'addName' in request.form and 'addPrice' in request.form and 'addCategory' in request.form:
+            item_name = request.form['addName']
+            item_price = request.form['addPrice']
+            item_category = request.form['addCategory']
             cursor.execute('INSERT INTO menu (itemName, itemPrice, itemCategory) VALUES (%s, %s, %s)', (item_name, item_price, item_category))
             mysql.connection.commit()
             return redirect(url_for('menu'))
-    if request.method == 'POST' and 'update-item' in request.form:
-            item_name = request.form['update-item']
-            item_price = request.form['update-price']
+        if request.method == 'POST' and 'editName' in request.form and 'updatePrice' in request.form:
+            item_name = request.form['editName']
+            item_price = request.form['updatePrice']
             cursor.execute('UPDATE Menu SET itemPrice = %s WHERE itemName = %s', (item_price, item_name))
             mysql.connection.commit()
             return redirect(url_for('menu'))
-    if request.method == 'POST' and 'special-name' in request.form and 'special-percent' in request.form and 'special-duration' in request.form:
-            special_name = request.form['special-name']
-            special_percent = request.form['special-percent']
-
-            cursor.execute('UPDATE Menu SET itemPrice = itemPrice * %s WHERE itemName = %s', (1 - float(special_percent) / 100, special_name),)
+        if request.method == 'POST' and 'specialName' in request.form and 'specialPercent' in request.form:
+            special_name = request.form['specialName']
+            special_percent = request.form['specialPercent']
+            special_percent = int(special_percent) / 100 + 1
+            cursor.execute('UPDATE Menu SET itemPrice = itemPrice * %s WHERE itemName = %s', (special_percent, special_name))
 
             return redirect(url_for('menu'))
-    
     if is_admin():
         return render_template('menu.html', menu_items=menu_items, is_admin=is_admin())
     else:
@@ -173,6 +173,7 @@ def logout():
     session.pop('loggedin', None)
     session.pop('id', None)
     session.pop('username', None)
+    session.pop('isAdmin', None)
     return redirect(url_for('login'))
 
 @app.route('/profile', methods=['GET', 'POST'])
@@ -222,10 +223,10 @@ def admprofile():
         return redirect(url_for('login'))
 
 def is_admin():
-    cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-    cursor.execute('SELECT Username, Email FROM UserInfo WHERE isAdmin = FALSE')
-    account = cursor.fetchone()
-    return account['isAdmin']
+    if 'isAdmin' in session:
+        return True
+    else:
+        return False
 
 # Running the Flask app
 if __name__ == '__main__':
